# openAPI.yaml
openapi: 3.0.0
info:
  title: Event & Ticket Management API
  version: 1.0.0-oas3-oas3.1
servers:
  - url: http://localhost:5000
paths:
  /user/signup:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
  /user/validate-email:
    post:
      summary: Send email verification token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              required:
                - email
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad Request
  /user/login:
    post:
      summary: Verify user credentials and get a token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                  _id:
                    type: string
                  name:
                    type: string
                  surname:
                    type: string
        '400':
          description: Bad Request
  /user/verify/{id}:
    get:
      summary: Verify user account with the provided token
      parameters:
        - name: id
          in: path
          description: Verification token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /user/current-user:
    get:
      summary: Get information about the current user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                  _id:
                    type: string
                  name:
                    type: string
                  surname:
                    type: string
        '400':
          description: Bad Request
  /user/reset-password:
    post:
      summary: Send a reset password email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              required:
                - email
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad Request
  /user/new-password:
    post:
      summary: Set a new password for the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                password:
                  type: string
              required:
                - id
                - password
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad Request
        '404':
          description: Not Found
...
components:
  schemas:
    User:
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        verified:
          type: boolean
          default: false
        resetToken:
          type: string
        expireToken:
          type: string
          format: date-time
        my_events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        created_events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        attendedActivities:
          type: array
          items:
            $ref: '#/components/schemas/EventActivity'
        checkedInActivities:
          type: array
          items:
            $ref: '#/components/schemas/EventActivity'
        created_at:
          type: string
          format: date-time
    Event:
      type: object
      properties:
        name:
          type: string
        short_description:
          type: string
        description:
          type: string
        creator:
          type: string
          format: uuid
        bannerURL:
          type: string
        location:
          type: string
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        category:
          type: string
        price:
          type: number
          format: double
        program:
          type: object
        attendees:
          type: array
          items:
            type: string
            format: uuid
        checked_in_attendees:
          type: array
          items:
            type: string
            format: uuid
        activities:
          type: array
          items:
            $ref: '#/components/schemas/EventActivity'
        created_at:
          type: string
          format: date-time
    EventActivity:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        time:
          type: string
          format: date-time
        location:
          type: string
        event:
          type: string
          format: uuid
        creator:
          type: string
          format: uuid
        attendees:
          type: array
          items:
            type: string
            format: uuid
        checkedInAttendees:
          type: array
          items:
            type: string
            format: uuid
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []